upd fronted and backend

+++ backend/internal/domain/model/leaderboard.go
@@ -4,5 +4,5 @@ import "github.com/google/uuid"
 
 type LeaderboardRow struct {
 	ProfileID uuid.UUID
-	Points    int
+	Points    float64
 }
